Refactors the organization of the code
Re adjusted the way Travis engage the pytest  and correceted fundamental pathing issue causint the coverage and pytests failures.
