+++ .travis.yml
@@ -9,10 +9,11 @@ install:
 # command to run tests
 
 script:
-  - pytest
+  - cd "Python"
+  - python -m pytest
   # - nosetests --with-coverage
 
 after_success:
-  - coverage run ./Python/Coverage/coverage_setup.py  
+  - coverage run ./Coverage/coverage_setup.py  
   - coverage report -m
-  - coveralls --rcfile=./Python/.coveragerc
+  - coveralls --rcfile=./.coveragerc
