+++ Python/LORIS/timepoint.py
@@ -1,12 +1,10 @@
 import sys
 import os
 import json
-import argparse
-import getpass
 import logging
 from dotenv import load_dotenv
-from LORIS_query import number_extraction, getCNBP, putCNBP, login
-from LORIS_candidates import check_DCCID, checkDCCIDExist
+from LORIS.query import number_extraction, getCNBP, putCNBP, login
+from LORIS.candidates import check_DCCID, checkDCCIDExist
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
