+++ Python/tests/test_DICOM.py
@@ -1,11 +1,9 @@
-import pydicom
-import sys
 import os
-import argparse
-import getpass
+import sys
 import logging
 import unittest
-from ..DICOM import DICOM_RequireDecompression, DICOM_validator, DICOM_retrieveMRN, DICOM_computeScanAge, DICOM_anonymizer, DICOM_updateElement, DICOM_retrieveElements
+
+from DICOM.DICOM import DICOM_RequireDecompression, DICOM_validator, DICOM_retrieveMRN, DICOM_computeScanAge, DICOM_anonymizer, DICOM_updateElement, DICOM_retrieveElements
 from pydicom.data import get_testdata_files
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
