+++ Python/tests/test_Intermediate_Server.py
@@ -1,8 +1,8 @@
-from ..LORIS_query import login
-from ..LocalDB_create_CNBP import create_localDB_CNBP
-from ..LocalDB_schema import *
-from ..LocalDB_query import create_entry, update_entry
-from ..Intermediate_LORIS_LocalDB import findTimePointUpdateDatabase
+from LORIS.query import login
+from LocalDB.create_CNBP import create_localDB_CNBP
+from LocalDB.schema import *
+from LocalDB.query import create_entry, update_entry
+from Intermediate_LORIS_LocalDB import findTimePointUpdateDatabase
 
 
 def test_updateLocalTimepoint():
