+++ Python/tests/test_LocalDBCreate.py
@@ -2,9 +2,9 @@ from pathlib import Path
 import logging
 import os
 import sys
-from ..LocalDB_create_CNBP import LocalDBCreate, create_localDB_CNBP
-from ..LocalDB_query import check_header
-from ..LocalDB_schema import *
+from LocalDB.create_CNBP import LocalDBCreate, create_localDB_CNBP
+from LocalDB.query import check_header
+from LocalDB.schema import *
 
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
