+++ Python/tests/test_LocalDBQuery.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import logging
 import os
 import sys
-from ..LocalDB_query import check_value, update_entry, create_entry
-from ..LocalDB_create_CNBP import create_localDB_CNBP
+from LocalDB.query import check_value, update_entry, create_entry
+from LocalDB.create_CNBP import create_localDB_CNBP
 
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
