+++ Python/tests/test_file_ops.py
@@ -1,5 +1,5 @@
 import os
-from ..file_operation import recursive_list_files, copy_files_to_flat_folder
+from file_operation import recursive_list_files, copy_files_to_flat_folder
 
 
 def test_recursive_load():
